Update setRep.js
made the code more reliable, there was an issure when it came to finding the score value.
It should be working now but there seems to be a problem when it comes to updating the calue, and it insted only updates the previous one. This will be cheked later when -rep is fixed